+++ gpu-video/examples/decode.rs
@@ -1,4 +1,4 @@
-#[cfg(vulkan)]
+#[cfg(supported)]
 fn main() {
     use std::io::Write;
 
@@ -68,9 +68,9 @@ fn main() {
     writer_thread_handle.join().unwrap();
 }
 
-#[cfg(not(vulkan))]
+#[cfg(not(supported))]
 fn main() {
     println!(
-        "This crate doesn't work on your operating system, because it does not support vulkan"
+        "This crate doesn't work on your operating system, because it does not support vulkan or video toolbox"
     );
 }