+++ gpu-video/CHANGELOG.md
@@ -16,10 +16,12 @@
       - etc...
   - Removed `vk-validation` and `vk-api-dump` features. Instead you can enable them when creating a `VideoInstance` or `wgpu::Instance`
 - Updated `wgpu` to `30.0.0` ([#2111](https://github.com/software-mansion/smelter/pull/2111) by @noituri)
+- `WgpuTexturesDecoderH264` now returns frames via a callback, the same way `BytesDecoderH264` does ([#2186](https://github.com/software-mansion/smelter/pull/2186) by @jerzywilczek)
 
 ### ✨ New features
 - Added VUI support to H.265 encoder ([#1995](https://github.com/software-mansion/smelter/pull/1995) by @noituri)
 - Made decoders work asynchronously. Bytes decoder now returns frames via callback ([#2143](https://github.com/software-mansion/smelter/pull/2143) by @noituri)
+- Added an asynchronous H.264 decoder backed by VideoToolbox on macOS (by @jerzywilczek)
 
 ### 🐛 Bug fixes
 - Request only the transfer and compute queues that are actually used when creating the device. ([#2223](https://github.com/software-mansion/smelter/pull/2223) by @jerzywilczek)